USB configuration + debug
Started throwing in some USB configurations, it required clock tree changes to the peripherals already used. From 72MHz to 48MHz which is still very fast for my purposes,  so I had to adjust the prescalers for TIM6 and TIM7. I also added PA7 (Trace2) as GPIO output to debug the timings. So far the 200ms target is respected between 1024 samples and all the elaborations (the tasks in waterfall) require 12,5ms to be carried out.